requirements.txt
pandas
numpy
requests
tk
openpyxl
python-dotenv